New methods added

+++ src/User.java
@@ -11,10 +11,12 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 public class User {
+    private String defaultUserName = "__defaultUserName__";
     private String username;
     private String password;
+    private String credentials;
     private int id;
-    private String defaultUserName = "__defaultUserName__";
+    
     
     public User() {
         this.username = defaultUserName;
@@ -28,39 +30,50 @@ public class User {
         //en caso de necesitarlo se haría
     }
     
-    // Tries to authenticate the user. Returns true if the user was found in the DB and the password is correct
+    // Returns true if the user was found in the DB and sets the user credentials
     // Returns false otherwise
+    private boolean userExists (String username) {
+        String fileName = ".\\users.txt";
+        List<String> users = new ArrayList<>();
+        try (BufferedReader br = Files.newBufferedReader(Paths.get(fileName))) {
+            users = br.lines().filter(line -> line.startsWith(this.username + "|")).collect(Collectors.toList());            
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+        if (users.isEmpty()) {
+            System.out.println("User " + this.username + " does not exist.");
+            return false;
+        } else {
+            this.credentials = users.get(0);
+            return true;
+        }
+    }
+        // Tries to authenticate the user. 
     public boolean authUser() {
         if (this.username != "" && this.username != null) {
-            String fileName = ".\\users.txt";
-            List<String> users = new ArrayList<>();
-            try (BufferedReader br = Files.newBufferedReader(Paths.get(fileName))) {
-                users = br.lines().filter(line -> line.startsWith(this.username)).collect(Collectors.toList());            
-            } catch (IOException e) {
-                e.printStackTrace();
+            if (userExists(this.username)) {
+                
             }
-            if (users.isEmpty()) {
-                System.out.println("User " + this.username + " does not exist.");
-                return false;
-            }
-            return true;
         }
         return false;
     }
     
     // Adds the user to the DB - returns true if the user was added correctly
     public boolean setUser() {
-        String newUser = this.username + "|" + this.password;
-        try (FileWriter fw = new FileWriter(".\\users.txt", true);
-        BufferedWriter bw = new BufferedWriter(fw);
-        PrintWriter out = new PrintWriter(bw)) {
-            out.println(newUser);
-            fw.close();
-        } catch (IOException e) {
-            e.printStackTrace();
-            return false;
+        if (!userExists(this.username)) {
+            String newUser = this.username + "|" + this.password;
+            try (FileWriter fw = new FileWriter(".\\users.txt", true);
+            BufferedWriter bw = new BufferedWriter(fw);
+            PrintWriter out = new PrintWriter(bw)) {
+                out.println(newUser);
+                fw.close();
+            } catch (IOException e) {
+                e.printStackTrace();
+                return false;
+            }
+            return true;
         }
-        return true;
+        return false;
     }
     
     public String getName() {
